requirements.txt
# AI Travel Assistant Requirements

# Core Framework
streamlit>=1.28.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# OpenAI & Azure OpenAI
openai>=1.6.0

# Vector Databases
pinecone>=3.0.0
chromadb>=0.4.0  # Optional: Use scripts/install_chromadb.bat if installation fails
scikit-learn>=1.3.0  # For cosine similarity in fallback mode

# LangChain Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Text-to-Speech
gTTS>=2.4.0

# JSON Processing
# json is built-in Python module

# File I/O
# io is built-in Python module

# Operating System
# os is built-in Python module

# Logging
# logging is built-in Python module